release controller: split cluster choosing from scheduling
As we're moving to scheduling releases in each cluster, it will become
useful to split choosing which clusters to schedule a release on from
the actual scheduling, as choosing the clusters can only happen once,
while we'll schedule the release once per cluster.
