+++ pkg/controller/release/release_controller_test.go
@@ -1440,28 +1440,6 @@ func (f *fixture) expectTrafficNotReady(relpair releaseInfoPair, targetStep, ach
 	f.expectedEvents = []string{}
 }
 
-func TestControllerComputeTargetClustersAndCreateAssociatedObjects(t *testing.T) {
-	namespace := "test-namespace"
-	app := buildApplication(namespace, "test-app")
-	cluster := buildCluster("minikube")
-
-	f := newFixture(t, app.DeepCopy(), cluster.DeepCopy())
-	f.cycles = 1
-	contenderName := "test-contender"
-	var replicaCount int32 = 1
-	contender := f.buildContender(namespace, contenderName, replicaCount)
-	delete(contender.release.Annotations, shipper.ReleaseClustersAnnotation)
-
-	f.addObjects(
-		contender.release.DeepCopy(),
-	)
-
-	f.expectReleaseScheduled(contender.release, []*shipper.Cluster{cluster})
-	f.expectAssociatedObjectsCreated(contender.release, []*shipper.Cluster{cluster})
-
-	f.run()
-}
-
 func TestContenderReleasePhaseIsWaitingForCommandForInitialStepState(t *testing.T) {
 	namespace := "test-namespace"
 	app := buildApplication(namespace, "test-app")
@@ -2662,121 +2640,6 @@ func TestWaitingOnIncumbentCapacityProducesNoPatches(t *testing.T) {
 		incumbent: incumbent,
 	}
 	f.expectCapacityNotReady(relpair, 1, 0, Incumbent, "minikube")
-	f.run()
-}
-
-func TestControllerChooseClusters(t *testing.T) {
-	namespace := "test-namespace"
-	app := buildApplication(namespace, "test-app")
-
-	clusterA := buildCluster("minikube-a")
-	clusterB := buildCluster("minikube-b")
-
-	f := newFixture(t, app.DeepCopy(), clusterA.DeepCopy(), clusterB.DeepCopy())
-	f.cycles = 1
-
-	contender := buildRelease()
-	delete(contender.Annotations, shipper.ReleaseClustersAnnotation)
-
-	replicaCount := int32(2)
-	contender.Spec.Environment.ClusterRequirements.Regions[0].Replicas = &replicaCount
-	f.addObjects(contender.DeepCopy())
-
-	expected := contender.DeepCopy()
-	expected.Annotations[shipper.ReleaseClustersAnnotation] = fmt.Sprintf("%s,%s", clusterA.Name, clusterB.Name)
-	condScheduled := releaseutil.NewReleaseCondition(shipper.ReleaseConditionTypeScheduled, corev1.ConditionTrue, "", "")
-	releaseutil.SetReleaseCondition(&expected.Status, *condScheduled)
-	condStrategyExecuted := releaseutil.NewReleaseCondition(shipper.ReleaseConditionTypeStrategyExecuted, corev1.ConditionTrue, "", "")
-	releaseutil.SetReleaseCondition(&expected.Status, *condStrategyExecuted)
-
-	f.actions = []kubetesting.Action{
-		kubetesting.NewUpdateAction(
-			shipper.SchemeGroupVersion.WithResource("releases"),
-			contender.GetNamespace(),
-			expected),
-	}
-
-	f.filter = f.filter.Extend(actionfilter{[]string{"update"}, []string{"releases"}})
-
-	relKey := fmt.Sprintf("%s/%s", contender.GetNamespace(), contender.GetName())
-	f.expectedEvents = append(f.expectedEvents,
-		fmt.Sprintf(
-			"Normal ClustersSelected Set clusters for \"%s\" to %s,%s",
-			relKey,
-			clusterA.Name,
-			clusterB.Name,
-		),
-		fmt.Sprintf(
-			"Normal ReleaseScheduled Created InstallationTarget \"%s\"",
-			relKey,
-		),
-		fmt.Sprintf(
-			"Normal ReleaseScheduled Created TrafficTarget \"%s\"",
-			relKey,
-		),
-		fmt.Sprintf(
-			"Normal ReleaseScheduled Created CapacityTarget \"%s\"",
-			relKey,
-		),
-		"Normal ReleaseConditionChanged [Scheduled False] -> [Scheduled True], [] -> [StrategyExecuted True]",
-	)
-
-	f.run()
-}
-
-func TestControllerChooseClustersSkipsUnschedulable(t *testing.T) {
-	namespace := "test-namespace"
-	app := buildApplication(namespace, "test-app")
-
-	clusterA := buildCluster("minikube-a")
-	clusterB := buildCluster("minikube-b")
-	clusterB.Spec.Scheduler.Unschedulable = true
-
-	f := newFixture(t, app.DeepCopy(), clusterA.DeepCopy(), clusterB.DeepCopy())
-	f.cycles = 1
-
-	contender := buildRelease()
-	f.addObjects(contender.DeepCopy())
-
-	expected := contender.DeepCopy()
-	expected.Annotations[shipper.ReleaseClustersAnnotation] = clusterA.Name
-
-	condScheduled := releaseutil.NewReleaseCondition(shipper.ReleaseConditionTypeScheduled, corev1.ConditionTrue, "", "")
-	releaseutil.SetReleaseCondition(&expected.Status, *condScheduled)
-	condStrategyExecuted := releaseutil.NewReleaseCondition(shipper.ReleaseConditionTypeStrategyExecuted, corev1.ConditionTrue, "", "")
-	releaseutil.SetReleaseCondition(&expected.Status, *condStrategyExecuted)
-
-	f.actions = []kubetesting.Action{
-		kubetesting.NewUpdateAction(
-			shipper.SchemeGroupVersion.WithResource("releases"),
-			contender.GetNamespace(),
-			expected),
-	}
-
-	f.filter = f.filter.Extend(actionfilter{[]string{"update"}, []string{"releases"}})
-
-	relKey := fmt.Sprintf("%s/%s", contender.GetNamespace(), contender.GetName())
-	f.expectedEvents = append(f.expectedEvents,
-		fmt.Sprintf(
-			"Normal ClustersSelected Set clusters for \"%s\" to %s",
-			relKey,
-			clusterA.Name,
-		),
-		fmt.Sprintf(
-			"Normal ReleaseScheduled Created InstallationTarget \"%s\"",
-			relKey,
-		),
-		fmt.Sprintf(
-			"Normal ReleaseScheduled Created TrafficTarget \"%s\"",
-			relKey,
-		),
-		fmt.Sprintf(
-			"Normal ReleaseScheduled Created CapacityTarget \"%s\"",
-			relKey,
-		),
-		"Normal ReleaseConditionChanged [Scheduled False] -> [Scheduled True], [] -> [StrategyExecuted True]",
-	)
-
 	f.run()
 }
