+++ pkg/controller/release/scheduler.go
@@ -1,17 +1,14 @@
 package release
 
 import (
-	"fmt"
 	"sort"
 	"strings"
 
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/client-go/tools/record"
 	helmchart "k8s.io/helm/pkg/proto/hapi/chart"
-	"k8s.io/klog"
 
 	shipper "github.com/bookingcom/shipper/pkg/apis/shipper/v1alpha1"
 	shipperchart "github.com/bookingcom/shipper/pkg/chart"
@@ -20,17 +17,14 @@ import (
 	listers "github.com/bookingcom/shipper/pkg/client/listers/shipper/v1alpha1"
 	"github.com/bookingcom/shipper/pkg/controller"
 	shippererrors "github.com/bookingcom/shipper/pkg/errors"
-	releaseutil "github.com/bookingcom/shipper/pkg/util/release"
 )
 
 type Scheduler struct {
 	clientset shipperclientset.Interface
 
-	clusterLister            listers.ClusterLister
 	installationTargetLister listers.InstallationTargetLister
 	trafficTargetLister      listers.TrafficTargetLister
 	capacityTargetLister     listers.CapacityTargetLister
-	rolloutBlockLister       listers.RolloutBlockLister
 
 	chartFetcher shipperrepo.ChartFetcher
 
@@ -39,22 +33,18 @@ type Scheduler struct {
 
 func NewScheduler(
 	clientset shipperclientset.Interface,
-	clusterLister listers.ClusterLister,
 	installationTargerLister listers.InstallationTargetLister,
 	capacityTargetLister listers.CapacityTargetLister,
 	trafficTargetLister listers.TrafficTargetLister,
-	rolloutBlockLister listers.RolloutBlockLister,
 	chartFetcher shipperrepo.ChartFetcher,
 	recorder record.EventRecorder,
 ) *Scheduler {
 	return &Scheduler{
 		clientset: clientset,
 
-		clusterLister:            clusterLister,
 		installationTargetLister: installationTargerLister,
 		trafficTargetLister:      trafficTargetLister,
 		capacityTargetLister:     capacityTargetLister,
-		rolloutBlockLister:       rolloutBlockLister,
 
 		chartFetcher: chartFetcher,
 
@@ -62,51 +52,7 @@ func NewScheduler(
 	}
 }
 
-func (s *Scheduler) ChooseClusters(rel *shipper.Release) (*shipper.Release, error) {
-	metaKey := controller.MetaKey(rel)
-	if releaseHasClusters(rel) {
-		return nil, shippererrors.NewUnrecoverableError(fmt.Errorf("release %q has already been assigned to clusters", metaKey))
-	}
-	klog.V(4).Infof("Choosing clusters for release %q", metaKey)
-
-	selector := labels.Everything()
-	allClusters, err := s.clusterLister.List(selector)
-	if err != nil {
-		return nil, shippererrors.NewKubeclientListError(
-			shipper.SchemeGroupVersion.WithKind("Cluster"),
-			"", selector, err)
-	}
-
-	selectedClusters, err := computeTargetClusters(rel, allClusters)
-	if err != nil {
-		return nil, err
-	}
-	setReleaseClusters(rel, selectedClusters)
-
-	return rel, nil
-}
-
 func (s *Scheduler) ScheduleRelease(rel *shipper.Release) (*releaseInfo, error) {
-	metaKey := controller.MetaKey(rel)
-	klog.V(4).Infof("Processing release %q", metaKey)
-	defer klog.V(4).Infof("Finished processing %q", metaKey)
-
-	if !releaseHasClusters(rel) {
-		var err error
-		rel, err = s.ChooseClusters(rel)
-		if err != nil {
-			return nil, err
-		}
-		s.recorder.Eventf(
-			rel,
-			corev1.EventTypeNormal,
-			"ClustersSelected",
-			"Set clusters for %q to %v",
-			controller.MetaKey(rel),
-			rel.Annotations[shipper.ReleaseClustersAnnotation],
-		)
-	}
-
 	replicaCount, err := s.fetchChartAndExtractReplicaCount(rel)
 	if err != nil {
 		return nil, err
@@ -139,10 +85,6 @@ func (s *Scheduler) ScheduleRelease(rel *shipper.Release) (*releaseInfo, error) 
 		trafficTarget:      tt,
 		capacityTarget:     ct,
 	}, nil
-}
-
-func releaseHasClusters(rel *shipper.Release) bool {
-	return len(rel.Annotations[shipper.ReleaseClustersAnnotation]) > 0
 }
 
 func stringSliceEqual(arr1, arr2 []string) bool {
@@ -246,9 +188,6 @@ func (s *Scheduler) CreateOrUpdateInstallationTarget(rel *shipper.Release) (*shi
 	it, err := s.installationTargetLister.InstallationTargets(rel.GetNamespace()).Get(rel.GetName())
 	if err != nil {
 		if !errors.IsNotFound(err) {
-			klog.Errorf("Failed to get InstallationTarget %q from lister interface: %s",
-				controller.MetaKey(rel),
-				err)
 			return nil, err
 		}
 		it := &shipper.InstallationTarget{
@@ -289,15 +228,9 @@ func (s *Scheduler) CreateOrUpdateInstallationTarget(rel *shipper.Release) (*shi
 	}
 
 	if !installationTargetClustersMatch(it, clusters) {
-		klog.V(4).Infof("Updating InstallationTarget %q clusters to %s",
-			controller.MetaKey(it),
-			strings.Join(clusters, ","))
 		setInstallationTargetClusters(it, clusters)
 		updIt, err := s.clientset.ShipperV1alpha1().InstallationTargets(rel.GetNamespace()).Update(it)
 		if err != nil {
-			klog.Errorf("Failed to update InstallationTarget %q clusters: %s",
-				controller.MetaKey(it),
-				err)
 			return nil, err
 		}
 		s.recorder.Eventf(
@@ -319,9 +252,6 @@ func (s *Scheduler) CreateOrUpdateCapacityTarget(rel *shipper.Release, totalRepl
 	ct, err := s.capacityTargetLister.CapacityTargets(rel.GetNamespace()).Get(rel.GetName())
 	if err != nil {
 		if !errors.IsNotFound(err) {
-			klog.Errorf("Failed to get CapacityTarget %q from lister interface: %s",
-				controller.MetaKey(rel),
-				err)
 			return nil, err
 		}
 		ct := &shipper.CapacityTarget{
@@ -357,15 +287,9 @@ func (s *Scheduler) CreateOrUpdateCapacityTarget(rel *shipper.Release, totalRepl
 	}
 
 	if !capacityTargetClustersMatch(ct, clusters) {
-		klog.V(4).Infof("Updating CapacityTarget %q clusters to %s",
-			controller.MetaKey(ct),
-			strings.Join(clusters, ","))
 		setCapacityTargetClusters(ct, clusters, totalReplicaCount)
 		updCt, err := s.clientset.ShipperV1alpha1().CapacityTargets(rel.GetNamespace()).Update(ct)
 		if err != nil {
-			klog.Errorf("Failed to update CapacityTarget %q clusters: %s",
-				controller.MetaKey(ct),
-				err)
 			return nil, err
 		}
 		s.recorder.Eventf(
@@ -387,9 +311,6 @@ func (s *Scheduler) CreateOrUpdateTrafficTarget(rel *shipper.Release) (*shipper.
 	tt, err := s.trafficTargetLister.TrafficTargets(rel.GetNamespace()).Get(rel.GetName())
 	if err != nil {
 		if !errors.IsNotFound(err) {
-			klog.Errorf("Failed to get TrafficTarget %q from lister interface: %s",
-				controller.MetaKey(rel),
-				err)
 			return nil, err
 		}
 		tt := &shipper.TrafficTarget{
@@ -425,15 +346,9 @@ func (s *Scheduler) CreateOrUpdateTrafficTarget(rel *shipper.Release) (*shipper.
 	}
 
 	if !trafficTargetClustersMatch(tt, clusters) {
-		klog.V(4).Infof("Updating TrafficTarget %q clusters to %s",
-			controller.MetaKey(tt),
-			strings.Join(clusters, ","))
 		setTrafficTargetClusters(tt, clusters)
 		updTt, err := s.clientset.ShipperV1alpha1().TrafficTargets(rel.GetNamespace()).Update(tt)
 		if err != nil {
-			klog.Errorf("Failed to update TrafficTarget %q clusters: %s",
-				controller.MetaKey(tt),
-				err)
 			return nil, err
 		}
 		s.recorder.Eventf(
@@ -449,128 +364,6 @@ func (s *Scheduler) CreateOrUpdateTrafficTarget(rel *shipper.Release) (*shipper.
 	return tt, nil
 }
 
-// computeTargetClusters picks out the clusters from the given list which match
-// the release's clusterRequirements.
-func computeTargetClusters(rel *shipper.Release, clusterList []*shipper.Cluster) ([]*shipper.Cluster, error) {
-	regionSpecs := rel.Spec.Environment.ClusterRequirements.Regions
-	requiredCapabilities := rel.Spec.Environment.ClusterRequirements.Capabilities
-	capableClustersByRegion := map[string][]*shipper.Cluster{}
-	regionReplicas := map[string]int{}
-
-	if len(regionSpecs) == 0 {
-		return nil, shippererrors.NewNoRegionsSpecifiedError()
-	}
-
-	app, err := releaseutil.ApplicationNameForRelease(rel)
-	if err != nil {
-		return nil, err
-	}
-
-	err = validateClusterRequirements(rel.Spec.Environment.ClusterRequirements)
-	if err != nil {
-		return nil, err
-	}
-
-	prefList := buildPrefList(app, clusterList)
-	// This algo could probably build up hashes instead of doing linear searches,
-	// but these data sets are so tiny (1-20 items) that it'd only be useful for
-	// readability.
-	for _, region := range regionSpecs {
-		capableClustersByRegion[region.Name] = []*shipper.Cluster{}
-		if region.Replicas == nil {
-			regionReplicas[region.Name] = 1
-		} else {
-			regionReplicas[region.Name] = int(*region.Replicas)
-		}
-
-		matchedRegion := 0
-		for _, cluster := range prefList {
-			if cluster.Spec.Scheduler.Unschedulable {
-				continue
-			}
-
-			if cluster.Spec.Region == region.Name {
-				matchedRegion++
-				capabilityMatch := 0
-				for _, requiredCapability := range requiredCapabilities {
-					for _, providedCapability := range cluster.Spec.Capabilities {
-						if requiredCapability == providedCapability {
-							capabilityMatch++
-							break
-						}
-					}
-				}
-
-				if capabilityMatch == len(requiredCapabilities) {
-					capableClustersByRegion[region.Name] = append(capableClustersByRegion[region.Name], cluster)
-				}
-			}
-		}
-		if regionReplicas[region.Name] > matchedRegion {
-			return nil, shippererrors.NewNotEnoughClustersInRegionError(region.Name, regionReplicas[region.Name], matchedRegion)
-		}
-	}
-
-	resClusters := make([]*shipper.Cluster, 0)
-	for region, clusters := range capableClustersByRegion {
-		if regionReplicas[region] > len(clusters) {
-			return nil, shippererrors.NewNotEnoughCapableClustersInRegionError(
-				region,
-				requiredCapabilities,
-				regionReplicas[region],
-				len(clusters),
-			)
-		}
-
-		//NOTE(btyler): this assumes we do not have duplicate cluster names. For the
-		//moment cluster objects are cluster scoped; if they become namespace scoped
-		//and releases can somehow be scheduled to clusters from multiple namespaces,
-		//this assumption will be wrong.
-		for _, cluster := range clusters {
-			if regionReplicas[region] > 0 {
-				regionReplicas[region]--
-				resClusters = append(resClusters, cluster)
-			}
-		}
-	}
-
-	sort.Slice(resClusters, func(i, j int) bool {
-		return resClusters[i].Name < resClusters[j].Name
-	})
-
-	return resClusters, nil
-}
-
-func validateClusterRequirements(requirements shipper.ClusterRequirements) error {
-	// Ensure capability uniqueness. Erroring instead of de-duping in order to
-	// avoid second-guessing by operators about how Shipper might treat repeated
-	// listings of the same capability.
-	seenCapabilities := map[string]struct{}{}
-	for _, capability := range requirements.Capabilities {
-		_, ok := seenCapabilities[capability]
-		if ok {
-			return shippererrors.NewDuplicateCapabilityRequirementError(capability)
-		}
-		seenCapabilities[capability] = struct{}{}
-	}
-
-	return nil
-}
-
-func setReleaseClusters(rel *shipper.Release, clusters []*shipper.Cluster) {
-	clusterNames := make([]string, 0, len(clusters))
-	memo := make(map[string]struct{})
-	for _, cluster := range clusters {
-		if _, ok := memo[cluster.Name]; ok {
-			continue
-		}
-		clusterNames = append(clusterNames, cluster.Name)
-		memo[cluster.Name] = struct{}{}
-	}
-	sort.Strings(clusterNames)
-	rel.Annotations[shipper.ReleaseClustersAnnotation] = strings.Join(clusterNames, ",")
-}
-
 func (s *Scheduler) fetchChartAndExtractReplicaCount(rel *shipper.Release) (int32, error) {
 	chart, err := s.chartFetcher(&rel.Spec.Environment.Chart)
 	if err != nil {
@@ -581,8 +374,6 @@ func (s *Scheduler) fetchChartAndExtractReplicaCount(rel *shipper.Release) (int3
 	if err != nil {
 		return 0, err
 	}
-
-	klog.V(4).Infof("Extracted %d replicas from release %q", replicas, controller.MetaKey(rel))
 
 	return int32(replicas), nil
 }
